get_log interface rollback

+++ include/kbase/scheduler.h
@@ -48,7 +48,7 @@ namespace crx
          * @root_dir 设置日志的根目录，将在根目录基础上按照年/月/日逐级构造子目录
          * @max_size 文件切割大小(单位为MB)
          */
-        log get_log(const char *prefix, const char *root_dir = "log_files", int max_size = 10);
+        log get_log(const char *prefix, const char *root_dir = "log_files", int64_t max_size = 10);
 
         //[单例]
         sigctl get_sigctl();
